+++ lihil/__init__.py
@@ -14,12 +14,12 @@ from .constant import status as status
 # from .interface import AppState as AppState
 from .interface import HTML as HTML
 from .interface import MISSING as MISSING
+from .interface import SSE as SSE
 from .interface import Empty as Empty
+from .interface import EventStream as EventStream
 from .interface import Json as Json
 from .interface import Payload as Payload
 from .interface import Stream as Stream
-from .interface import SSE as SSE
-from .interface import EventStream as EventStream
 from .interface import Text as Text
 from .lihil import Lihil as Lihil
 from .local_client import LocalClient as LocalClient
@@ -36,5 +36,5 @@ from .websocket import WebSocketRoute as WebSocketRoute
 
 # from .server.runner import run as run
 
-VERSION = "0.2.29"
+VERSION = "0.3.30"
 __version__ = VERSION